refactor(execution): split executor per HandlerIntent variant
execute_one() and simulate() were one giant 3-arm match each, totalling
~450 LOC of arm bodies plus their helpers. Each arm and its helpers
now lives in a sibling file:

- execution/link.rs   — Link arm (deploy via double symlink), with
                        ancestor-cycle detection and conflict handling
- execution/stage.rs  — Stage arm (datastore staging), with auto-chmod
                        and dry-run reporting helpers for the path
                        handler bin/ tree
- execution/run.rs    — Run arm (sentinel-gated command execution)

execution/mod.rs keeps the Executor struct, the public execute() entry
point, and the dispatchers that delegate to per-intent methods. The
per-intent implementations live in additional `impl<'a> Executor<'a>`
blocks in their respective files; visibility is `pub(super)` so nothing
outside `execution` can call them directly.

Note: the execute_one match is on Intent variants (Link/Stage/Run),
not handler types — `path` chmod sits inside Stage, `install` and
`homebrew` both flow through Run. The split mirrors the actual Intent
shape rather than the handler taxonomy.
